PR tidy up
